refactor mutation: update ticket handler on empty string
