+++ graphql/resolvers/ticket/mutations.js
@@ -9,10 +9,17 @@ const mutations = {
     return await toDelete.delete()
   },
   async updateTicket (root, { id, input }) {
+    const toUpdate = await Ticket.findOrFail(id)
+    let data = {}
+      !input.title ? data.title = toUpdate.title : data.title = input.title
+      !input.description ? data.description = toUpdate.description : data.description = input.description
+      !input.category_id ? data.category_id = toUpdate.category_id : data.category_id = input.category_id
+      !input.status ? data.status = toUpdate.status : data.status = input.status
+
     await Ticket
       .query()
       .where('id', id)
-      .update(input)
+      .update(data)
 
     return await Ticket.findOrFail(id)
   },
